moved tests into exporter_test and slots_test
